+ 5.3.0.0
+ Optimized Ultima SDK API bootstrapping.
+ Housekeeping updates for old syntax patterns to .NET 4.8 minimum standard.
+ Improved compatibility and support for ServUO Publish 58.
+ Deprecated support for RunUO.

+ Renamed README file to README.md
+ Renamed LICENSE file to LICENSE.md

+ Removed redundant VERSION file.
+ Removed redundant JsonReader and JsonWriter.
+ Removed redundant InstallationLocator feature.
+ Removed redundant PacketHandler API 6017 hooks.
+ Removed redundant Sector API extension methods.
+ Removed redundant SpawnZone API.
+ Removed redundant MySQL ODBC API.

+ Improved SuperGump API support for enhanced clients.
+ Improved SuperGump API support for dynamic-range scrollbars.
+ Improved SuperGump API support for fluid-layout controls.
+ Improved SuperGump API support for pagination controls.
+ Improved SuperGump API support for entity tool-tip controls.
+ Improved SuperGump API support for text and font controls.

+ Improved WebAPI compatibility by updating to a task-based request process.
+ Updated WebAPI to display unhandled requests in the console window while in DEBUG mode.

+ Optimized ObjectPool API and its implementations.
+ Extended ObjectPool API to support concurrent collections.

+ Added more characters to the UniGlyph table.
+ Extended UOFont API to allow multiple font file search paths.

+ Normalized the API for modules and services to maintain consistency.
+ Improved start-up sequence for modules and services with minor optimizations.

+ Overhauled TimeBoosts API and interface to improve usability and ease of access.
+ Extended support for TimeBoostTokens value randomization and dupe command compatibility.

+ Updated ExtendedOPL API to improve usability and ease of access.
+ Optimized ExtendedOPL API generation and decompiling.

+ Optimized EquipmentSets API set validation and detection.

+ Improved AutoPvP API scenario outcome and mission completion detection process.
+ Extended AutoPvP API support for persisting battle shortcut commands.

+ Updated AutoDonate API transaction processing and history interface.

+ Optimized AdvancedVendor API for buying and selling.
+ Extended AdvancedVendor API to support setting dynamic restock delays.

+ Extended SpellUtility API to include Mastery and Racial circles.

+ Normalized SkillBuffInfo, StatBuffInfo, and ResistBuffInfo API for consistency.

+ Added more specific flag combinations to the ExpansionFlags enum.

+ Extended SkillCodex API to allow overriding of critical functions for custom implementations.

+ Extended SeveredHead API to allow for a custom head creation hook to be assigned.

+ Extended HueDeeds API to allow for easier hooking of the deed creation event.
+ Extended NameDeeds API to allow for easier hooking of the deed creation event.

+ Extended DataStore API to improve usability and ease of access.

+ Extended Enum extension methods API with support for detecting attributes assigned to enum members.
+ Extended Enum extension methods API with support for retrieving absolute flag values.

+ Updated Serialization extension method API to a task-based process during async usage.

+ Optimized Clilocs API to improve memory unloading and argument parsing.

+ Fixed an issue that would cause RuneCodex entries to corrupt their names or descriptions.
+ Fixed an issue that would cause a LayeredIconDefinition to crash the client when displayed.
+ Fixed an issue that would cause AutoPvP battle broadcasts to be inconsistent in their usage.
+ Fixed an issue that would prevent items from being dropped from the cursor.
+ Fixed an issue that would prevent reflected static methods from being dynamically invoked.
+ Fixed an issue that would sometimes produce mangled output during a Json API export.
+ Fixed an issue that would prevent some stats from persisting and displaying correctly from the WebStats API.
+ Fixed an issue that would cause an enumerable extension method to ignore a buffer and iterate unsafely.
+ Fixed an issue that would cause an enumerable extension method to iterate in the wrong direction.
+ Fixed an issue that would cause an enumerable extension method to throw an erroneous out-of-range exception.

+++ Extensions/Server/ContainerExt.cs
@@ -48,7 +48,7 @@ namespace Server
 
 			total = container.FindItemsByType(type, true)
 							 .Where(i => i != null && !i.Deleted && i.TypeEquals(type, children) && (predicate == null || predicate(i)))
-							 .Aggregate(total, (c, i) => c + (long)i.Amount);
+							 .Aggregate(total, (c, i) => c + i.Amount);
 
 			return total >= amount;
 		}
